Read and Parse REDCap Data Dictionary

Reads a REDCap data dictionary (CSV format) and parses the "Choices" column to create a named list of value-label pairs for use in data recoding. By default, it loads an online dictionary in the RAGE redcap repository, or as a fallback, an internal dictionary included with the package, but users can specify a different dictionary file via the dictUrl argument.

Usage

read_and_parse_dict(
  dictUrl =
    "https://raw.githubusercontent.com/RAGE-toolkit/rage-redcap/main/data_dictionaries/RAGEredcap_DataDictionary.csv",
  fallbackPath = system.file("extdata", "RABVlab_DataDictionary.csv", package =
    "rabvRedcapProcessing")
)

Arguments

dictUrl

A URL to the data dictionary CSV file. Defaults to GitHub raw link.

fallbackPath

Local fallback path, system.file().

Value

A named list of named vectors, where each list element corresponds to a field name, and each named vector maps value codes to human-readable labels.

Details

Special handling is provided for the "country" field, allowing both country names and codes to be valid inputs by building a dual-mapped dictionary.

Examples

dicts <- read_and_parse_dict()
names(dicts)
#>  [1] "sample_tissuetype"       "sample_buffer"          
#>  [3] "country"                 "hmpcr_n405"             
#>  [5] "rtassay"                 "lateral_flow_test"      
#>  [7] "fat"                     "drit"                   
#>  [9] "ngs_platform"            "nanopore_platform"      
#> [11] "illumina_platform"       "ngs_prep"               
#> [13] "ngs_library"             "ngs_analysis_type"      
#> [15] "negative_control_sample"
dicts[["fat"]]
#>   Positive   Negative         NA 
#> "Positive" "Negative"       "NA"